break up files
