+++ column_names.py
@@ -1,0 +1,10 @@
+IX_ALT_PRONOUNCE: int = 2
+IX_PRONOUN: int = 3
+IX_VERB: int = 4
+IX_GENDER: int = 5
+IX_SYLLABARY: int = 6
+IX_PRONOUNCE: int = 7
+IX_ENGLISH: int = 8
+IX_INTRO_NOTE: int = 9
+IX_END_NOTE: int = 10
+IX_APP_FILE: int = 11
